requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
PyYAML==6.0.2
openai==1.51.0
python-dotenv==1.0.1
httpx==0.27.0
httpcore==1.0.5
